+++ NEWS.md
@@ -1,13 +1,34 @@
-Keeper 1.0.0
+Keeper 2.0.0
 ============
+
+This is a complete re-implementation of Keeper. 
 
 Changes:
 
-1. Using the [ellmer](https://ellmer.tidyverse.org/index.html) package to standardize communication with LLMs. Users can now call the `reviewCases()` function to have the LLM adjudicate cases.
+1. Changed `createKeeper()` to `generateKeeper()`, and made the following changes:
+
+    - The concept set input arguments have been grouped in a single `keeperConceptSets` argument. 
+    
+    - The output has changed from a human-readable table to a structured intermediary format.
+    
+    - Removed the 'comorbidities' category from both the input and the output, as this was ill-defined. 
+    
+    - The `databaseId` argument has been removed. Instead, the database name is extracted from the `cdm_source` table.
+
+2. Added the `convertKeeperToTable()` to convert the output of `generateKeeper()` to the original human-readable table format.
+
+3. Added the `reviewCases()` function that uses large language models (LLMs) to review Keeper profiles. It Uses the [ellmer](https://ellmer.tidyverse.org/index.html) package to standardize communication with LLMs.
+
+4. Added the `generateKeeperConceptSets()` function that uses LLMs, the OHDSI Vocabulary vector store, and Phoebe to suggest concept sets to be used as input for `generateKeeper()`.
+
+5. Added a Shiny app for human review of Keeper profiles. This can be launched using the `launchReviewerApp()` function. Experts can also deploy the Shiny app on a Shiny server with PostgreSQL backend.
+
+6. Added the `createSensitiveCohort()` function. This creates a cohort based on the Keeper input concept sets which should include everyone in the database who potentially has the phenotype. This cohort can be used to estimate sensitivity (in addition to positive predictive value) of a phenotype algorithm.
+
 
 Bugfixes:
 
-1. Avoiding incorrect person IDs due to integer overflow by usng strings instead.
+1. Avoiding incorrect person IDs due to integer overflow by using strings instead.
 
 
 Keeper 0.2.1
@@ -31,7 +52,7 @@ Keeper 0.1.0
 
 Changes:
 
-1. Change name from 'KEEPER' to 'Keeper', and `createKEEPER()` to `createKeeper()` for consistency with HADES.
+1. Change name from 'Keeper' to 'Keeper', and `createKeeper()` to `createKeeper()` for consistency with HADES.
 
 2. Removed `vocabularyDatabaseSchema` argument of `createKeeper()`, which wasn't used, and the vocabulary should be in the CDM schema anyway.